# [SPARK-44571][SQL] Eliminate the Join by combine multiple Aggregates