--- a/sql/catalyst/src/main/scala/org/apache/spark/sql/internal/SQLConf.scala
+++ b/sql/catalyst/src/main/scala/org/apache/spark/sql/internal/SQLConf.scala
@@ -4831,6 +4831,23 @@ object SQLConf {
     .booleanConf
     .createWithDefault(false)
 
+   val COMBINE_JOINED_AGGREGATES_ENABLED =
+    buildConf("spark.sql.optimizer.combineJoinedAggregates.enabled")
+      .doc("When true, we attempt to eliminate join by combine aggregates " +
+        "to reduce the scan times and avoid shuffle.")
+      .version("4.0.0")
+      .booleanConf
+      .createWithDefault(false)
+
+  val MAX_TREE_NODE_NUM_OF_PREDICATE =
+    buildConf("spark.sql.optimizer.combineJoinedAggregates.maxTreeNodeNumOfPredicate")
+      .doc("Maximum tree node number of predicate. If tree node number of predicate exceeds the" +
+        "limit, CombineJoinedAggregates will not merging the aggregates connected with join.")
+      .version("4.0.0")
+      .intConf
+      .checkValue(_ > 0, "The threshold of tree node numbers should be positive")
+      .createWithDefault(10)
+
   /**
    * Holds information about keys that have been deprecated.
    *
@@ -5754,6 +5771,11 @@ class SQLConf extends Serializable with Logging with SqlApiConf {
 
   def legacyRaiseErrorWithoutErrorClass: Boolean =
     getConf(SQLConf.LEGACY_RAISE_ERROR_WITHOUT_ERROR_CLASS)
+
+  def combineJoinedAggregatesEnabled: Boolean =
+    getConf(SQLConf.COMBINE_JOINED_AGGREGATES_ENABLED)
+
+  def maxTreeNodeNumOfPredicate: Int = getConf(SQLConf.MAX_TREE_NODE_NUM_OF_PREDICATE)
 
   def stackTracesInDataFrameContext: Int = getConf(SQLConf.STACK_TRACES_IN_DATAFRAME_CONTEXT)