… sql/catalyst/src/test/scala/org/apache/spark/sql/catalyst/optimizer/CombineJoinedAggregatesSuite.scala
… sql/core/src/test/scala/org/apache/spark/sql/DataFrameAggregateSuite.scala