@cloud-fan 2023-08-01
Instead of showing an example query, can you define the general form of joined aggregates that can be merged?