@peter-toth 2023-08-02
@beliefer, I've upodated my PR with comments to ellaborate on how it handles `Filter`s: https://github.com/apache/spark/pull/37630/files#diff-3d3aa853c51c01216a7f7307219544a48e8c14fabe1817850e58739208bc406aR351-R524